Move public ProposalObject methods into a separate module
